+++ crates/cargo-contract/src/cmd/printer.rs
@@ -1,13 +1,22 @@
 // Credit: https://github.com/wlezzar/jtab
-use std::collections::HashMap;
+use std::{
+    collections::HashMap,
+    fs::File,
+    io::{BufReader, Read},
+    path::Path,
+};
 
-use anyhow::bail;
+use anyhow::{bail, Error, Result};
 use prettytable::{Cell, format, Row, Table};
 use prettytable::format::{FormatBuilder, LinePosition, LineSeparator};
 use regex::Regex;
-use serde_json::Value;
+use serde_json::{
+    Value,
+};
 use yaml_rust::{Yaml, YamlEmitter};
 use yaml_rust::yaml::Hash;
+
+const BUILD_INFO_PATH: &str = "build_info.json";
 
 #[derive(Debug)]
 pub enum TableHeader {
@@ -224,4 +233,89 @@ impl Printer for PlainTextTablePrinter {
         table.printstd();
         Ok(())
     }
+}
+
+// Print a summary of all the smart contracts that have been built
+// that have been included in a file build_info.json in the project root.
+// If it is called whilst contract is being built then `write_new_build`
+// will be `true` and that contract will be added to or updated in build_info.json
+pub fn print_build_info(write_new_build: bool, output_json: bool,
+    contract_name: Option<&str>, contract_map: Option<HashMap<&str, &str>>) -> Result<(), Error> {
+    let exists_build_info_path = Path::new(BUILD_INFO_PATH).exists();
+    if !exists_build_info_path {
+        // println!("not existing path");
+        anyhow::bail!("Unable to print from or write to file that does not exist");
+    }
+    // println!("existing path");
+    // build_info.json exists, so update it with the data
+    let file_build_info = File::open(BUILD_INFO_PATH)?;
+    let mut buf_reader = BufReader::new(&file_build_info);
+    let mut contents = String::new();
+    buf_reader.read_to_string(&mut contents)?;
+    let mut build_info_json: Vec<HashMap<&str, &str>> =
+        serde_json::from_slice::<Vec<HashMap<&str, &str>>>(&contents.as_bytes())?;
+    // println!("build_info_json {:#?}", build_info_json);
+
+    if write_new_build == true {
+        let mut found = false;
+        for info in build_info_json.iter_mut() {
+            // replace existing build info with new contract info
+            // if the contract name already exists as a value in build_info.json
+            let c = match info.get(&"Contract") {
+                Some(c) => c,
+                None => "",
+            };
+            if c == contract_name.unwrap() {
+                found = true;
+                info.insert("Size", match contract_map.clone().unwrap().get(&"Size") {
+                    Some(s) => s,
+                    None => "",
+                });
+                info.insert("Metadata Path", match contract_map.clone().unwrap().get(&"Metadata Path") {
+                    Some(m) => m,
+                    None => "",
+                });
+            }
+        }
+        // if did not find an existing value in build_info_json to update
+        // then push the new value to the end
+        if found == false {
+            build_info_json.push(contract_map.clone().unwrap());
+        }
+        // write updated to file
+        serde_json::to_writer(&File::create("build_info.json")?,
+        &build_info_json.clone())?;
+    }
+    if output_json {
+        println!("{:#?}", &build_info_json);
+    } else {
+        // if they don't specify `--output-build-info-json` when running
+        // `cargo contract summary --output-build-info-json` then we will output tabular format
+        let build_info_json_value: Value = serde_json::to_value(&build_info_json).unwrap();
+        // println!("build_info_json_value {:#?}", &build_info_json_value);
+
+        let spec = vec!["Contract:Flipper:ddd".to_string()];
+        let colorize: Vec<_> = spec
+            .iter()
+            .map(ColorizeSpec::parse)
+            .collect::<Result<_, _>>()?;
+        // println!("colorize {:#?}", colorize);
+
+        // note: we actually don't need to provide headers because `infer_headers` infers the headers
+        // so it would still work if `given_headers` was `None`
+        let mut named_fields: Vec<String> = build_info_json[0].clone().into_keys().into_iter().map(|s| String::from(s)).collect::<Vec<String>>();
+        named_fields.sort_unstable();
+        // println!("named_fields {:#?}", named_fields);
+        // assert_eq!(named_fields, ["Contract", "Metadata Path", "Size"]);
+        let given_headers = TableHeader::NamedFields { fields: named_fields };
+
+        let table = JsonTable::new(Some(given_headers), &build_info_json_value);
+        // println!("table {:#?}", table);
+
+        // set to `PlainTextTableFormat::Default` or `PlainTextTableFormat::Markdown`
+        let format = PlainTextTableFormat::Default;
+        PlainTextTablePrinter::new(colorize, format).print(&table)?
+    }
+
+    Ok(())
 }
